feat(injector)!: Inject the Matomo tracker at module import
The Matomo tracker is now automatically injected in the application when the ngx-Matomo module is being imported. The configuration of the tracker is passed as a JSON object in order to specify the tracker JS script, URL, the different trackers URLs (more than one is now possible), and additional parameters driving the first actions of the tracker(s).
BREAKING CHANGE: since the injection is now automatic, the old way (0.x) of injecting the tracker manually is no longer compatible with this library.